refactor: move typed graph-op model to proximadb-graph-model leaf
PR 1 of the ORION extraction cascade (root-crate decomposition). Moves the
typed graph write-operation model — GraphOperation, NodeUpdate, EdgeUpdate
(+ NodeId/EdgeId aliases) — from src/storage/memtable/implementations/
graph_memtable.rs into the proximadb-graph-model foundation leaf, where their
payload types (Node, Edge, PropertyValue, EmbeddingVersion) already live.

graph_memtable.rs re-exports them (pub use proximadb_graph_model::{...}) so
every crate::storage::memtable::implementations::graph_memtable::* path —
notably the unified WAL's GraphOp variant (wal_operations.rs) and the ORION
graph engine — resolves unchanged. Type-compatible: root's crate::graph::
{Node,Edge,...} already resolve to this leaf via the src/graph/model.rs
re-export shim, and NodeId/EdgeId are transparent String aliases.

This unblocks the next cascade steps: a GraphWalPort trait (PR 2) can now name
GraphOperation from the leaf, and the ORION move (PR 3+) no longer forces a
cyclic root dependency through the typed graph-op model.

Pure relocation + re-export, no behavior change. (Note: the separate, flat
GraphOperation in proximadb-storage-transaction is a different transaction-DSL
type, intentionally untouched.)
